util, sorted list of commits in dir
Along with each commit is an epoch, and path of the repo.
'-T|--top' filtering is jankey (used as taken from
'find_repos_in_dir' (should we not be using it instead of duplicating
it?)). Ongoing is a reliable way to filter out 'not ours' repos.
Continue: output options for each commit.